+++ MK88Hydra/AppVersion.cs
@@ -2,7 +2,7 @@ namespace Hydra
 {
     internal static class AppVersion
     {
-        public const string Version = "1.0.0";
+        public const string Version = "1.0.1";
         public const string Guid = "com.mursisru.mk88hydra";
         public const string Name = "MK-88 Hydra";
     }